+++ R/nobs.RprobitB_fit.R
@@ -1,5 +1,0 @@
-#' @exportS3Method
-
-nobs.RprobitB_fit <- function(object, ...) {
-  sum(object$data$T)
-}